internal/provider: add handling cluster deleted status
fixes #63
